refactor(service): 本地 API 路由按领域拆分，聚合层只做注册
将 routes.ts 拆为 routes/ 子目录：system（whoami / version）、pr（列表 / 详情 / diff / 刷新 /
分类 / 评审写动作）、agent（Agent 操作），公共原语（Route 类型 / seg / NO_EVENT）归 shared；
routes/index 只做聚合注册与路径匹配、不含业务逻辑。与 CLI 命令树的领域划分对称。纯结构调整，
端点与行为不变。
